// frontend/src/ipc/types/AppStatus.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AppId } from "./AppId";

export type AppStatus = { "kind": "idle" } | { "kind": "starting", app: AppId, } | { "kind": "running", app: AppId, } | { "kind": "stopping" } | { "kind": "error", message: string, };