// frontend/src/ipc/types/AppMetadata.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AppId } from "./AppId";
import type { Direction } from "./Direction";
import type { RegulatoryClass } from "./RegulatoryClass";

export type AppMetadata = { id: AppId, name: string, direction: Direction, regulatory_class: RegulatoryClass, };